tree: normalize phandle managmement
A phandle of 0 or -1 is invalid, and we can't have it in
any output DTS.

We also don't need/want explicit properties in a node for
phandle management. That is done via the .phandle attribute
of LopperNode's.

We do need to print non-zero phandles, as numeric references
to them may have been created, and if we allow phandles to
be recreated/regenerated after a write -> FDT -> read, we
will break all those references.

This change ensures that properties are used as little as
possible, and that a phandle of 0 or -1 will never be
printed.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>