schema: complete path based typing
Some properties are different types depending on their context.

This previously worked due to heuristics and type guessing, but
now that we are typing inputs via scan we need to store more than
a single type for a given name.

We do this by creating a union of the identified types and keeing
the count of occurences of each type. If a property name is passed
and no path supplied, then the most common type is returned. If
a path is passed, you get the type of the property at that type.

To make the property path typing possible, we need to keep track
of the node path when parsing the dts. That was previously broken
due to label references in preprocessed files resetting/breaking
the path calculations. We make the scanning code smarter and adapt
to those types of structures.

Finall, we don't need to store path based typing for properties
that are found to have the same type. So once we've built the
path based typing, we optimized it and only keep the details
for properties that have different types througout the input.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>