schema: workaround mixed-type /bits/ properties
While mixing types of a single property name is not advised and
could be considered a semantic error, it does happen in some
inputs (which are force compiled or otherwise).

The issue isn't seen if ALL properties have a /bits/ directive
once it has started to be used, it is in a situation where there
is one property of that name with /bits/ and then other without.

When we detected a /bits/ -> no bits type property, treat it as
two different types so that our path based property type resolution
will kick in and we'll get closer to expected output in all cases.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>