schema/tree: add /bits/ 16 array and /bits/ 8 support
We previously were using binary data [00 01 02 03], etc
to handle 8 bit values when doing binary type guessing.

We have the ability to support /bits/ via the resolver,
so when /bits/ 8 or a /bits/ 16 array are detected we
generate a format hint in the schema and the tree
property resolver will use that information to improve
our output format.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>