lopper/tree: add trusted mode for node addtions
If a node is added from within the tree itself, we can
trust that the phandles and other references are valid
and there's no need to clear them.

This fixes node movment creating invalid phandles, as
the properties within the tree for phandles are stored
numerically and not symbolically.

If the symbol is stored in the future, we can return to
clearing all phandles on add and simplify the code.

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>