assists/subsystem: reserved_memory_expand: change to follow phandle list convention
Now that we expect user to define reserved memory only by inputs (YAML or SDT definitions),
the input for  reserved memory in YAML is list of strings.

For example:
    reserved-memory:
      - X
      - Y
where X and Y are node names.

The routine will simply look for the nodes in the reserved memory list

As YAML may define only start/size tuple, if that is the case then transform start/size
to 'reg' property in form of [ 0, start, 0, size ] that can be processed later.

the reserved-memory property for a domain will become list of phandles:

domainZ{
  reserved-memory = <phandleX>, <phandleY>;
};

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>