lopper/log: add additional logging levels
When verbose and print statements were elimated, we collapsed
about 5 logging levels into one. We need more granularity in
debug logging to allow for -v -v to not overwhelm the user.

Here we introduce some additional levels, expose them to the
callers via a type and a parameter to the _debug routine()

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>