lopper: allow .dtsi files to be joined with the SDT
.dtsi (overlays) are tagged as support files, but were
never allowed to be considered as available for merging
into the main system device tree.

To allow this we add .dtsi to the valid extensions.

This opens the door to having changes to the main SDT
in overlay files (subject to the standard dtc rules
of what can be in an overlay)

Signed-off-by: Bruce Ashfield <bruce.ashfield@amd.com>