gen_domain_dts: Add routine to return whether node should stay for OpenAMP cases
As there are multiple cases for OpenAMP that determine whether a node should stay
for a given domain - remove custom handling for this from gen-domain plugin
instead and have openamp domain return True for such cases.

This will ensure that as future cases come up in the future, this logic
does not balloon in gen-domain plugin.

This currently includes zephyr mbox, ipm nodes and Linux UIO nodes

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>